requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
asttokens==2.2.1
    # via stack-data
backcall==0.2.0
    # via ipython
beautifulsoup4==4.12.2
    # via -r requirements.in
certifi==2023.7.22
    # via requests
charset-normalizer==3.2.0
    # via requests
coverage==7.3.0
    # via -r requirements.in
decorator==5.1.1
    # via ipython
django==4.2.4
    # via
    #   -r requirements.in
    #   django-auto-prefetch
django-auto-prefetch==1.7.0
    # via -r requirements.in
executing==1.2.0
    # via stack-data
idna==3.4
    # via requests
ipython==8.14.0
    # via -r requirements.in
jedi==0.19.0
    # via ipython
matplotlib-inline==0.1.6
    # via ipython
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   tweepy
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
prompt-toolkit==3.0.39
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pygments==2.16.1
    # via ipython
python-dotenv==1.0.0
    # via -r requirements.in
requests==2.31.0
    # via
    #   -r requirements.in
    #   requests-oauthlib
    #   tweepy
requests-oauthlib==1.3.1
    # via tweepy
six==1.16.0
    # via asttokens
soupsieve==2.4.1
    # via beautifulsoup4
sqlparse==0.4.4
    # via django
stack-data==0.6.2
    # via ipython
traitlets==5.9.0
    # via
    #   ipython
    #   matplotlib-inline
tweepy==4.14.0
    # via -r requirements.in
unittest-parametrize==1.3.0
    # via -r requirements.in
urllib3==2.0.4
    # via requests
wcwidth==0.2.6
    # via prompt-toolkit
